third_party/WebKit/LayoutTests/bluetooth/requestDevice/same-device.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let devices = [];
  let push = device => devices.push(device);

  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: [heart_rate.alias]}]}))
    .then(push)
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: [heart_rate.name]}]}))
    .then(push)
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: [heart_rate.uuid]}]}))
    .then(push)
    .then(() => {
    assert_equals(devices[0], devices[1]);
    assert_equals(devices[1], devices[2]);
  });
}, 'Returned device should always be the same.');
</script>
